+++ src/algs/QuadMesher/JaalMoabConverter.cpp
@@ -155,7 +155,7 @@ JaalMoabConverter ::fromMOAB(iMesh_Instance imesh, Jaal::Mesh *m, iBase_EntitySe
      SimpleArray<iBase_EntityHandle> facenodes;
 
      jaalnode.clear();
-     for( int i = 0; i < numFaces; i++) {
+     for( int i = 0; i < (int) numFaces; i++) {
           iMesh_getEntAdj(imesh, tfaceHandles[i], iBase_VERTEX, ARRAY_INOUT(facenodes), &err);
           for (int j = 0; j < 3; j++)
                jaalnode[ facenodes[j] ] = NULL;
